Separate StarChartHandler into its own thing
